added slice package
