+++ slice/slice.go
@@ -1,0 +1,13 @@
+package slice
+
+func UniqueSlice[S ~[]E, E comparable](s S) S {
+	m := map[E]struct{}{}
+
+	var out S
+	for _, item := range s {
+		if _, ok := m[item]; !ok {
+			out = append(out, item)
+		}
+	}
+	return out
+}
